refactor : Clinic BottomSheet 수정 (#96)
