+++ src/features/clinic/hook/useSelectAlarmTime.ts
@@ -1,0 +1,13 @@
+import { useState } from 'react'
+
+import { ALARM_TIME } from '@/constants'
+
+export const useSelectAlarmTime = () => {
+  const [selectedTime, setSelectedTime] = useState<string>(ALARM_TIME[5])
+
+  const handleClickTime = (time: string) => {
+    setSelectedTime(time)
+  }
+
+  return { selectedTime, handleClickTime }
+}
